ref: unify formatter functions
